Add valid_move? function to Board class
Add a function called valid_move? to the Board class that returns true if
the move given (from a certain position to a new position) is a valid move.
Also add a line to the move_piece function that sets the moved variable to
true for a pawn after it has already moved.