Add check? function to Board class and fix 2 bugs
Create a check? function that returns true if the given color's king is in
a checked position. Also fix a bug in the no_pieces_in_between_diagonal?
function so that the function works for NW direction. Also change line 53
to use the moves instance variable of piece instead of the possible_moves
instance function.