Add #checkmate? and modify #valid_move? and #find_king?
Add #checkmate? function that when given a color, returns true if the
color's king is checkmated. Write tests to test @checkmate?.

Also modify #valid_move? to fix a bug that previously allowed the pawn
to move forward to take an enemy piece.

Modify #find_king? to return the king object instead of the king's location.