Fix bug in #checkmate? and refactor #check? in Board class
Fix a bug in #checkmate? so that the function now checks to see if a
friendly piece can kill the enemy piece that is checking the king.

Also refactor #check? by using a helper function #pieces_checking_king,
which returns an array of all pieces that are checking the king in question.